Adding title tag to main template
the title tag is useful for search engines. also it will be displayed in the title bar or/and in the tab of the browsers. Our project now can be also be better read by bots and potential customer. thats why the title tag was evented almost 10 years ago!

Nehme mal an so erwartest du die Commits Message oder flurischt? ;)